validating for delete

+++ database/ticket.py
@@ -17,13 +17,13 @@ def create_ticket(ticket: dict):
     try:
         data = ticket["person"]
         person_response = requests.post(url = PERSON_ENDPOINT, json= data, headers=headers)
-        #print(ticket["person"])
-        #print(r.status_code)
+        print("person sortId: " , person_response.json()["sortId"])
         if person_response.status_code == 200:
             table.put_item(Item=ticket)
             return ticket
         else :
             return JSONResponse(content="Person not valid", status_code=person_response.status_code)
+             
     except ClientError as e:
         return JSONResponse(content=e.response["Error"], status_code=500)
 
@@ -33,7 +33,7 @@ def get_ticket(id: str):
         response = table.query(
             KeyConditionExpression=Key("id").eq(id)
         )
-        if response["Count"] > 1:
+        if response["Count"] > 0:
             return response["Items"]
         else:
             raise HTTPException(status_code=404, detail="Item not found")
@@ -44,7 +44,7 @@ def get_ticket(id: str):
 def get_tickets():
     try:
         response = table.scan(
-            Limit=5,
+            Limit=10,
             AttributesToGet=["id", "concert", "sector", "person", "date", "quantity", "created_at"]
         )
         return response["Items"]
@@ -52,16 +52,21 @@ def get_tickets():
         return JSONResponse(content=e.response["Error"], status_code=500)
 
 
-def delete_ticket(ticket: dict):
+def delete_ticket(id: str, created_at: str):
     try:
-        response = table.delete_item(
-            Key={
-                "id": ticket["id"],
-                "created_at": ticket["created_at"]
-            }
+        item = table.query(
+            KeyConditionExpression=Key("id").eq(id)
         )
-        print(response)
-        return response
+        if item["Count"] > 0:
+            response = table.delete_item(
+            Key={
+                "id": id,
+                "created_at": created_at
+            }
+            )
+            return response
+        else:
+            raise HTTPException(status_code=404, detail="Item not found")
     except ClientError as e:
         return JSONResponse(content=e.response["Error"], status_code=500)
 
@@ -73,14 +78,16 @@ def update_ticket(ticket: dict):
                 "id": ticket["id"],
                 "created_at": ticket["created_at"]
             },
-            UpdateExpression="SET #quantity = :quantity, #date = :date",
+            UpdateExpression="SET #quantity = :quantity, #date = :date, #sector = :sector",
             ExpressionAttributeValues={
                 ":quantity": ticket["quantity"],
-                ":date": ticket["date"]
+                ":date": ticket["date"],
+                ":sector": ticket["sector"]
             },
             ExpressionAttributeNames={
                 "#quantity": "quantity",
-                "#date": "date"
+                "#date": "date",
+                "#sector": "sector"
             }
         )
         return response
